app/requirements.txt
numpy==1.26.4
scikit-learn==1.4.2
annotated-types==0.6.0
anyio==4.3.0
audioread==3.0.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.14.0
fsspec==2024.3.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numba==0.59.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
platformdirs==4.2.1
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
sympy==1.12
threadpoolctl==3.5.0
torch==2.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
textrank @ git+https://github.com/lovit/textrank.git@578c7fbbb22b24561b1bd4bab9a9ecc2dc9515e1